autenticacao aula07

+++ Docker/start-project/app/Http/Controllers/EixoController.php
@@ -1,7 +1,6 @@
 <?php
 
 namespace App\Http\Controllers;
-
 use App\Models\Eixo;
 use Dompdf\Dompdf;
 use Illuminate\Http\Request;
@@ -13,8 +12,9 @@ class EixoController extends Controller
      */
     public function index()
     {
+        $this->authorize('index', Eixo::class);
         $data = Eixo::all();//all metodo que da o select *
-        //dd($data);
+        //Storage::disk('local')->put('example.txt', 'Contents');
         return view('eixo.index', compact('data'));
     }
 
@@ -23,6 +23,7 @@ class EixoController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Eixo::class);
         return view('eixo.create');
     }
 
@@ -31,6 +32,7 @@ class EixoController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('store', Eixo::class);
         if($request->hasFile('documento')){
 
             $eixo = new Eixo();
@@ -55,6 +57,8 @@ class EixoController extends Controller
      */
     public function show(string $id)
     {
+        
+        $this->authorize('show', Eixo::class);
         $eixo = Eixo::find($id);
         if(isset($eixo)){
             return view('eixo.show', compact(['eixo']));
@@ -67,6 +71,7 @@ class EixoController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('edit', Eixo::class);
         $eixo = Eixo::find($id);
         if(isset($eixo)){
             return view('eixo.edit', compact(['eixo']));
@@ -79,6 +84,7 @@ class EixoController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('update', Eixo::class);
         $eixo = Eixo::find($id);
         if(isset($eixo)){
             $eixo->nome = $request->nome;//setando
@@ -94,7 +100,7 @@ class EixoController extends Controller
      */
     public function destroy(string $id)
     {
-        $eixo = Eixo::find($id);
+        $this->authorize('destroy', Eixo::class);        $eixo = Eixo::find($id);
         if(isset($eixo)){
             $eixo->delete();
             return redirect()->route('eixo.index');
